force initialize order
